+++ src/blocks/logo/__img/logo__img.css
@@ -7,5 +7,6 @@
   transform: rotate(225deg);
   margin: -10px 0 0 auto;
   position: relative;
-  z-index: 5;
+  /* Я не поняла, зачем здесь был z-index 5, совсем убирать не стала, чтобы не отвалилось */
+  z-index: 1;
 }
