+++ src/blocks/logo/logo.css
@@ -1,3 +1,7 @@
 @import url(./__img/logo__img.css);
 @import url(./__container/logo__container.css);
 @import url(./__title/logo__title.css);
+
+.logo {
+  text-decoration: none;
+}
